Move reduce_all_dims to common
Also remove unused includes
