+++ common/CMakeLists.txt
@@ -2,7 +2,7 @@
 # contributors (https://github.com/scipp)
 set(TARGET_NAME "scipp-common")
 set(INC_FILES include/scipp/common/index.h include/scipp/common/overloaded.h
-              include/scipp/common/span.h
+              include/scipp/common/reduction.h include/scipp/common/span.h
 )
 
 add_library(${TARGET_NAME} INTERFACE)
